Move inital tagless code to Technique.Evaluator stub
This started with work embeddeing an encoding in the initial or final
tagless style after Oleg Kiselyov, but evolved to a pure abstract syntax
tree as a parsing target and feed to pretty printing. Move the evaluator
stubs out to a scratch pad file and leaving Technique.Language for the
AST.
